[#5] 코드 정리 및 재배치
