+++ src/handle/edit_files.rs
@@ -1,0 +1,25 @@
+use convert_case::{Case, Casing};
+use std::fs;
+
+use crate::utils::get_config_file_or_warn;
+
+pub fn add_usecase_interface_to_domain_file_if_not_exists(domain: String) {
+    let config_file = get_config_file_or_warn();
+
+    let domain_file_path = config_file.domain_dir.join(domain.clone() + ".go");
+
+    let usecase_interface_name =
+        domain.as_str().to_case(Case::Pascal) + &config_file.usecase_interface_suffix;
+
+    let mut domain_file_content = fs::read_to_string(&domain_file_path).unwrap();
+
+    if !domain_file_content.contains(&usecase_interface_name) {
+        domain_file_content.push_str("\n\n");
+
+        domain_file_content.push_str(format!(r#"// {usecase_interface_name}\n"#).as_str());
+        domain_file_content
+            .push_str(format!(r#"type {usecase_interface_name} interface {{}}"#).as_str());
+
+        fs::write(domain_file_path, domain_file_content).unwrap();
+    }
+}
