+++ src/handle/mod.rs
@@ -2,3 +2,4 @@ pub mod api;
 pub mod create_files;
 pub mod lookup;
 pub mod render;
+pub mod edit_files;
